Finally! It all began when I decided to display the frame length. So I found out that I can use FreeType to draw characters. There are some tutorials on how to use it, so it was easy. But, as I looked through the code, I found it too profligate to have a separate texture for every glyph, and to generate vertex data for every glyph whenever I needed to draw it. So I quickly came up with an idea to generate a one-dimensional texture with all the glyphs bitmaps there. Also, I wanted to generate vertex data only once, cause position, scale, and color could be passed via uniforms.
So I began. The first problem was that I didn't know that if your GPU allow us to have a maximum 16K two-dimensional texture, it means that you also can have only 16K one-dimensional texture. And 16K wasn't sufficient enough for my 625360 bytes. But I only found it after I had added debugging functionality. I even successfully printed my 1D texture int a .txt file, a-la ASCII graphic! So, I switched to 2D texture.
I didn't want to use extra space, so, I tried to factorize my 625360, and it came as 80 x 7817. Good. So I wrote a function to factorize my glyph atlas. I still had some problems with texture mapping, there was a shift, and I didn't know why, but the letters were visible and recognizable. I decided to reduce the size of a glyph, from 128 to 64, just to check. Boom! Total atlas's size was 159364 it's 4 * 39841, and 39841 is a prime number. So I just decided to take a square root and round up. There will be some extra free space in the texture atlas, but I was ok with it.
But I still had that annoying shift. You know, I put a screenshot in a folder. But after a few days, I solved it by adding glPixelStorei(GL_UNPACK_ALIGNMENT, 1);
I thought so. But now I don't know whether it was the cause or not. Because if I comment it, everything still works fine. Also, a found out that integers are not interpolated when passed from vertex to fragment shader.